+++ src/test/resources/function_with_declare.sql
@@ -4,11 +4,13 @@ AS
 $$
 DECLARE
     v_result INTEGER;
+    v_result_2 INTEGER;
 BEGIN
     v_result = 1;
+    v_result_2 = 1;
     -- Some code
     -- Some code
     -- Some code
-    RETURN v_result;
+    RETURN v_result + v_result_2;
 END;
 $$LANGUAGE plpgsql;