+++ src/test/resources/function_without_declare.sql
@@ -9,6 +9,3 @@ BEGIN
     RETURN 0;
 END;
 $$LANGUAGE plpgsql;
-
-
-