+++ src/test/resources/function_without_declare_with_comments.sql
@@ -10,6 +10,3 @@ BEGIN
     RETURN 0;
 END;
 $$LANGUAGE plpgsql;
-
-
-