+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [Unreleased]
 ### Bugfixes
-- Incorrect line position [#86]()
+- Incorrect line position [#86](https://github.com/ng-galien/idea-plpgdebugger/issues/86)
 
 ## [222.2.0]
 ### Features